gb::joypad

Struct Buttons

pub struct Buttons(/* private fields */);

The eight buttons, set where held.

The d-pad occupies the low nibble and the action buttons the high one, which is the layout GBDK’s J_ constants use.

Implementations

impl Buttons

pub const fn new() -> Self

Creates a new default initialized bitfield.

pub const fn from_bits(bits: u8) -> Self

Convert from bits.

pub const fn into_bits(self) -> u8

Convert into bits.

pub const fn right(&self) -> bool

Right on the d-pad.

Bits: 0..1

pub const fn with_right_checked(self, value: bool) -> Result<Self, ()>

Right on the d-pad.

Bits: 0..1

pub const fn with_right(self, value: bool) -> Self

Right on the d-pad.

Bits: 0..1

pub const fn set_right(&mut self, value: bool)

Right on the d-pad.

Bits: 0..1

pub const fn set_right_checked(&mut self, value: bool) -> Result<(), ()>

Right on the d-pad.

Bits: 0..1

pub const fn left(&self) -> bool

Left on the d-pad.

Bits: 1..2

pub const fn with_left_checked(self, value: bool) -> Result<Self, ()>

Left on the d-pad.

Bits: 1..2

pub const fn with_left(self, value: bool) -> Self

Left on the d-pad.

Bits: 1..2

pub const fn set_left(&mut self, value: bool)

Left on the d-pad.

Bits: 1..2

pub const fn set_left_checked(&mut self, value: bool) -> Result<(), ()>

Left on the d-pad.

Bits: 1..2

pub const fn up(&self) -> bool

Up on the d-pad.

Bits: 2..3

pub const fn with_up_checked(self, value: bool) -> Result<Self, ()>

Up on the d-pad.

Bits: 2..3

pub const fn with_up(self, value: bool) -> Self

Up on the d-pad.

Bits: 2..3

pub const fn set_up(&mut self, value: bool)

Up on the d-pad.

Bits: 2..3

pub const fn set_up_checked(&mut self, value: bool) -> Result<(), ()>

Up on the d-pad.

Bits: 2..3

pub const fn down(&self) -> bool

Down on the d-pad.

Bits: 3..4

pub const fn with_down_checked(self, value: bool) -> Result<Self, ()>

Down on the d-pad.

Bits: 3..4

pub const fn with_down(self, value: bool) -> Self

Down on the d-pad.

Bits: 3..4

pub const fn set_down(&mut self, value: bool)

Down on the d-pad.

Bits: 3..4

pub const fn set_down_checked(&mut self, value: bool) -> Result<(), ()>

Down on the d-pad.

Bits: 3..4

pub const fn a(&self) -> bool

The A button.

Bits: 4..5

pub const fn with_a_checked(self, value: bool) -> Result<Self, ()>

The A button.

Bits: 4..5

pub const fn with_a(self, value: bool) -> Self

The A button.

Bits: 4..5

pub const fn set_a(&mut self, value: bool)

The A button.

Bits: 4..5

pub const fn set_a_checked(&mut self, value: bool) -> Result<(), ()>

The A button.

Bits: 4..5

pub const fn b(&self) -> bool

The B button.

Bits: 5..6

pub const fn with_b_checked(self, value: bool) -> Result<Self, ()>

The B button.

Bits: 5..6

pub const fn with_b(self, value: bool) -> Self

The B button.

Bits: 5..6

pub const fn set_b(&mut self, value: bool)

The B button.

Bits: 5..6

pub const fn set_b_checked(&mut self, value: bool) -> Result<(), ()>

The B button.

Bits: 5..6

pub const fn select(&self) -> bool

The Select button.

Bits: 6..7

pub const fn with_select_checked(self, value: bool) -> Result<Self, ()>

The Select button.

Bits: 6..7

pub const fn with_select(self, value: bool) -> Self

The Select button.

Bits: 6..7

pub const fn set_select(&mut self, value: bool)

The Select button.

Bits: 6..7

pub const fn set_select_checked(&mut self, value: bool) -> Result<(), ()>

The Select button.

Bits: 6..7

pub const fn start(&self) -> bool

The Start button.

Bits: 7..8

pub const fn with_start_checked(self, value: bool) -> Result<Self, ()>

The Start button.

Bits: 7..8

pub const fn with_start(self, value: bool) -> Self

The Start button.

Bits: 7..8

pub const fn set_start(&mut self, value: bool)

The Start button.

Bits: 7..8

pub const fn set_start_checked(&mut self, value: bool) -> Result<(), ()>

The Start button.

Bits: 7..8

impl Buttons

pub const fn x(self) -> DPadX

Where the d-pad points left or right.

pub const fn y(self) -> DPadY

Where the d-pad points up or down.